% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_child.R
\name{sjsR_renderChild}
\alias{sjsR_renderChild}
\title{Compile child Rmd document into html character vector}
\usage{
sjsR_renderChild(child_rmd, td = tempdir(), ...)
}
\arguments{
\item{child_rmd}{path of child Rmd to compile}

\item{td}{temporary directory to use for compilation}

\item{...}{arguments to pass to rmarkdown::render}
}
\value{
A character vector of length 1 containing the compiled html
}
\description{
Compile child Rmd document into html character vector
}
